Hunter extension to find email addresses
Hunter for Chrome lets you find immediately who to contact when you visit a website.
If you already know the name the person you would like to contact, type it in the search field. The Email Finder will return the found email address along with a confidence score and sources.